+++ lib/tab/utils.js
@@ -32,6 +32,7 @@ function getTabInfo(jpTab){
         return null;
     }
 }
+exports.getTabInfo = getTabInfo;
 
 // Below code is based on adhacker, taken from http://forums.mozillazine.org/viewtopic.php?f=19&p=6335275
